+++ services/MemberService/src/main/java/api/service/MemberService/dto/login/LoginResponseDto.java
@@ -1,5 +1,6 @@
-package api.service.MemberService.dto;
+package api.service.MemberService.dto.login;
 
+import api.service.MemberService.dto.member.MemberDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -7,11 +8,12 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
-@NoArgsConstructor
 @AllArgsConstructor
-public class MemberResponseDto {
+@NoArgsConstructor
+public class LoginResponseDto {
 
     private int statusCode;
     private String message;
+    private String accessToken;
     private MemberDto memberInfo;
 }
